update add_new_voucher page
