+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.10.4+42",
+  "version": "0.10.4+46",
   "versionName": "RC1",
   "private": false,
   "type": "commonjs",
