+++ src/pages/users/accounting/add_new_voucher.tsx
@@ -3,20 +3,20 @@ import Head from 'next/head';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ILocale } from '@/interfaces/locale';
-import NewVoucherForm from '@/components/voucher/new_voucher_form';
+import NewVoucherFormRC2 from '@/components/voucher/new_voucher_form_rc2';
 import Layout from '@/components/beta/layout/layout';
 import { ISUNFA_ROUTE } from '@/constants/url';
-import { ICertificateUI } from '@/interfaces/certificate';
 import { useAccountingCtx } from '@/contexts/accounting_context';
+import { IInvoiceRC2UI } from '@/interfaces/invoice_rc2';
 // import { useUserCtx } from '@/contexts/user_context'; // Deprecated: (20250603 - Liz) 移除檢視者透過 URL 直接進入頁面的權限阻擋
 // import { TeamRole } from '@/interfaces/team'; // Deprecated: (20250603 - Liz) 移除檢視者透過 URL 直接進入頁面的權限阻擋
 
-const AddNewVoucherPage: React.FC = () => {
+const AddNewVoucherRC2Page: React.FC = () => {
   const { t } = useTranslation('common');
   // const { teamRole } = useUserCtx();
 
-  const [selectedCertificates, setSelectedCertificates] = useState<{
-    [id: string]: ICertificateUI;
+  const [selectedInvoices, setSelectedInvoices] = useState<{
+    [id: string]: IInvoiceRC2UI;
   }>({});
 
   const { clearReverseListHandler } = useAccountingCtx();
@@ -25,10 +25,10 @@ const AddNewVoucherPage: React.FC = () => {
     // Info: (20250305 - Julian) 進入此頁面時，清除 reverseList
     clearReverseListHandler();
 
-    const storedCertificates = localStorage.getItem('selectedCertificates');
-    if (storedCertificates) {
-      setSelectedCertificates(JSON.parse(storedCertificates));
-      localStorage.removeItem('selectedCertificates');
+    const storedInvoices = localStorage.getItem('selectedInvoices');
+    if (storedInvoices) {
+      setSelectedInvoices(JSON.parse(storedInvoices));
+      localStorage.removeItem('selectedInvoices');
     }
   }, []);
 
@@ -56,7 +56,7 @@ const AddNewVoucherPage: React.FC = () => {
         pageTitle={t('journal:ADD_NEW_VOUCHER.PAGE_TITLE')}
         goBackUrl={ISUNFA_ROUTE.BETA_VOUCHER_LIST}
       >
-        <NewVoucherForm selectedData={selectedCertificates} />
+        <NewVoucherFormRC2 selectedData={selectedInvoices} />
       </Layout>
     </>
   );
@@ -84,4 +84,4 @@ const getStaticPropsFunction = async ({ locale }: ILocale) => ({
 
 export const getStaticProps = getStaticPropsFunction;
 
-export default AddNewVoucherPage;
+export default AddNewVoucherRC2Page;
